pyproject.toml
[project]
name = "pdf-to-audio"
version = "0.1.0"
description = "A Streamlit application that converts PDFs to audio using LLM processing and TTS"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.32.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "elevenlabs>=0.3.0",
    "google-generativeai>=0.3.0",
    "google-cloud-aiplatform>=1.38.0",
    "pytest>=8.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.2",
    "reportlab>=4.1.0"
]
authors = [
    {name = "Your Name"}
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-v"

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
    "pytest-cov"
]

[tool.hatch.build.targets.wheel]
packages = ["src"] 